Refactor: Split WorkflowEditorPage into smaller components
Extract ReactFlow configuration and node management.
